feat: add iOS listening shader button
